# Speed stopping criterion computation by directly using unlist